+++ CHANGELOG.md
@@ -58,6 +58,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed the CLAY NPC
 - removed legacy material support
 ### Fixes
+- event priority for block objective
 - linebreaks in strings
 - notify:1 for block objective did not work
 - asynchronous database access for objectives